Rename ArchComponents to AndroidLifecycle.
Summary:
* Rename com.uber.autodispose.android.archcomponents package to com.uber.autodispose.android.lifecycle
* Rename AndroidLifecycleScopeProvider to just AndroidLifecycle

Resolves #91
