+++ autodispose-android-archcomponents/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="com.uber.autodispose.android.androidlifecycle"/>
+<manifest package="com.uber.autodispose.android.lifecycle"/>
